frontend/requirements.txt
scikit-learn~=1.0.2
numpy==1.24.4
pandas~=1.4.2
nltk~=3.7
seaborn~=0.11.2
matplotlib~=3.5.1
pyyaml == 6.0.1
wordcloud~=1.8.2.2
streamlit~=1.23.1